x86: Add config options to assign versions in the PE-COFF header
This commit adds the new config options to allow the user to modify the
following fields in the PE-COFF header.

UINT16 MajorOperatingSystemVersion
UINT16 MinorOperatingSystemVersion
UINT16 MajorImageVersion
UINT16 MinorImageVersion

Those fields are mainly for the executables or libraries in Windows NT
or higher to specify the minimum supported Windows version and the
version of the image itself.

Given the fact that those fields are ignored in UEFI, we can reuse those
fields for other purposes.

For SUSE Linux Enterprise, we will use them for "Security Version".

The idea of "Security Version" is to maintain a white list of the version
of the latest known secure kernel and to show a warning when loading an
insecure kernel.

First, we redefine those fields in the PE-COFF header:

* Signer Name: MajorOperatingSystemVersion + MinorOperatingSystemVersion
* Distro Version: MajorImageVersion
* Security Version: MinorImageVersion

The "Signer Name" is treated as a 4-bytes char array, while Distro
Version and Security Version are still UINT16.

For example, we can use this to represent the SLE-12-SP2 kernel with
security version 0:

* Signer Name: 'S''U''S''E'
* Distro Version: 0x1202
* Security Version: 0

Then, we will keep the white list in SVList, a UEFI BS+NV variable. The
elements in the list are stored in this way:

  [Signer 1][Distro version 1 ][Security version 1 ][Distro version 2][Security version 2]
  [Signer 2][Distro version 1'][Security version 1']
  ...

SVList will be mirrored into SVListRT, a RT volatile variable.
By design, the security versions in SVList only can be incremented
monotonically or a warning will show up to notify the user.

When grub2 uses shim protocol to verify the kernel, shim will match the
signer and distro version. Only the kernel with the higher or equal
security version can be booted.

There are a few ways to update SVList:

1. When SVList doesn't exist, shim create SVList based on the first boot
   kernel.
2. When booting a kernel with a higher security version, shim updates
   the existing security version automatically.
3. The RT+NV variable, SVListNew, is created and can be verified with
   the signature (signed by the built-in key or MOK) in SVListSig (RT+NV).
   As the design, shim only updates the higher security version or add
   the security versions with the new signer or new distro.
4. When the kernel is rejected, a prompt will show to ask the user if
   shim should lower the security version or add the new signer/distro.

To make reinstallation easier, shim will reset SVList when it's loaded
from a USB stick, DVD, PXE, or HTTP Boot.

If the distro wants to keep the list up-to-date, a system service can be
setup to check SVListRT and create SVListNew when necessary.

The other parts of Security Version implementations can be found in my
github repo:
* shim: https://github.com/lcp/shim/tree/svlist
* mokutil: https://github.com/lcp/mokutil/tree/svlist

For more details of Security Version, check the following page:
https://wiki.microfocus.net/index.php/Desktop_Preload/Taiwan/security_version

Cc: Vojtech Pavlik <vojtech@suse.cz>
Cc: Joey Lee <jlee@suse.com>
Cc: Michal Marek <mmarek@suse.com>
Cc: Matt Fleming <mfleming@suse.de>
Signed-off-by: Gary Lin <glin@suse.com>
Tested-by: Joey Lee <jlee@suse.com>